refactor: UserAnswer Dto 패키지 변경
