+++ back/src/main/java/com/cocktailpick/back/cocktail/dto/UserPreferenceAnswer.java
@@ -1,4 +1,4 @@
-package com.cocktailpick.back.cocktail.domain;
+package com.cocktailpick.back.cocktail.dto;
 
 public enum UserPreferenceAnswer {
 	YES, NO, SOSO;
